Split 'server' into redis utils and setup
